fix vscode extension file support
