+++ vscode-extension/package.json
@@ -97,6 +97,9 @@
         "id": "kotlin"
       },
       {
+        "id": "c"
+      },
+      {
         "id": "haskell"
       }
     ]
@@ -105,7 +108,7 @@
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
     "watch": "tsc -watch -p ./",
-    "test": "tsc -p ./ && node out/extractFindings.test.js"
+    "test": "tsc -p ./ && node out/extractFindings.test.js && node out/supportedFiles.test.js"
   },
   "devDependencies": {
     "@types/node": "^18.0.0",
